+++ test/benchmark/matrix.sol
@@ -38,23 +38,23 @@ contract MAT {
 		return rtn[a][a];
 	}
 	function mat64_mul(uint64 t) public pure returns(uint64) {
-		uint a = uint(0);
-		uint b = uint(1);
+		uint a = uint64(0);
+		uint b = uint64(1);
 		uint64[2][2] memory base;
-		base[a][a] = 1;
-		base[a][b] = 1;
-		base[b][a] = 1;
-		base[b][b] = 0;
+		base[a][a] = uint64(1);
+		base[a][b] = uint64(1);
+		base[b][a] = uint64(1);
+		base[b][b] = uint64(0);
 		uint64[2][2] memory rtn;
-		rtn[a][a] = 1;
-		rtn[b][b] = 1;
-		rtn[b][a] = 0;
-		rtn[a][b] = 0;
+		rtn[a][a] = uint64(1);
+		rtn[b][b] = uint64(1);
+		rtn[b][a] = uint64(0);
+		rtn[a][b] = uint64(0);
 		uint64[2][2] memory tmp;
 		for(uint cnt = uint(0); cnt < t; cnt += 1) {
 			for(uint i = uint(0); i < 2; i += 1) {
 				for(uint j = uint(0); j < 2; j += 1) {
-					tmp[i][j] = 0;
+					tmp[i][j] = uint64(0);
 				}
 			}
 			for(uint i = uint(0); i < 2; i += 1) {
